+++ src/main/java/ee/sk/smartid/rest/dao/SessionStatus.java
@@ -93,10 +93,14 @@ public class SessionStatus implements Serializable {
   }
 
   /**
-   * IP address of the device running the App.
-   * Present only for subscribed RPs and when available (e.g. not present in case state is TIMEOUT).
+   * IP-address of the device running the App.
    *
-   * @return IP address of the device running Smart-id app (or null if not returned)
+   * Present only if withShareMdClientIpAddress() was specified with the request
+   * Also, the RelyingParty must be subscribed for the service.
+   * Also, the data must be available (e.g. not present in case state is TIMEOUT).
+   * @see <a href="https://github.com/SK-EID/smart-id-documentation#238-mobile-device-ip-sharing">Mobile Device IP sharing</a>
+   *
+   * @return IP address of the device running Smart-ID app (or null if not returned)
    */
   public String getDeviceIpAddress() {
     return deviceIpAddress;